backend/requirements.txt
torch>=2.6.0
torchvision
torchaudio
librosa>=0.10.0
numpy>=1.23.5,<2
scipy
whisper-openai>=20231117
transformers>=4.49.0
accelerate>=1.1.1
moviepy==1.0.3
opencv-python>=4.9.0.80
imageio[ffmpeg]
imageio-ffmpeg
diffusers>=0.31.0
omegaconf
pandas
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pydub>=0.25.1
soundfile>=0.12.1
essentia>=2.1b6.dev1034
audioread>=3.0.0
